Added logging and moved moved database init.
 Changes to be committed:
	modified:   TeaTimer/App.xaml.cs
	modified:   TeaTimer/AppShell.xaml.cs
	modified:   TeaTimer/MauiProgram.cs
